Add support for --test-groups to wpt
This adds a --test-groups command line argument which points at a JSON
file containing lists of tests divided into explicit groups like:

{"/dom": ["/dom/historical.html", ...],
 "/dom/events/": [...]}

This is for situations where the division of tests into groups is
performed by an external process (in the case of gecko: by the
decision task).

Group names must be a path prefix, as this metadata is reused as the
test "scope" which is passed down into the output and can be used by
automatic metadata update for per-group properties like the LSAN allow
list.

--test-groups is incompatible with --run-by-dir but composes with
passing an explicit include list by running the intersection of the
supplied tests.

Differential Revision: https://phabricator.services.mozilla.com/D75175

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1634554
gecko-commit: 71aa576ada8bb74ddde1f213b2f6a46cada25f7f
gecko-integration-branch: autoland
gecko-reviewers: ahal, egao